+++ VDA-StorageNode-Registry/contracts/facets/VDAStorageNodeManagementFacet.sol
@@ -308,21 +308,21 @@ contract VDAStorageNodeManagementFacet is IStorageNodeManagement {
   /**
     * @dev see { IStorageNodeManagement }
     */
-  function isRegisteredName(string calldata name) external view returns(bool) {
+  function isRegisteredNodeName(string calldata name) external view returns(bool) {
     return LibStorageNode.nodeStorage()._nameNodeId[name] != 0;
   }
 
   /**
     * @dev see { IStorageNodeManagement }
     */
-  function isRegisteredAddress(address didAddress) external view returns(bool) {
+  function isRegisteredNodeAddress(address didAddress) external view returns(bool) {
     return LibStorageNode.nodeStorage()._didNodeId[didAddress] != 0;
   }
 
   /**
     * @dev see { IStorageNodeManagement }
     */
-  function isRegisteredEndpoint(string calldata endpointUri) external view returns(bool) {
+  function isRegisteredNodeEndpoint(string calldata endpointUri) external view returns(bool) {
     return LibStorageNode.nodeStorage()._endpointNodeId[endpointUri] != 0;
   }